mcp_tools/web_search/pyproject.toml
[project]
name = "web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.6.0",
    "tavily-python>=0.7.0",
]
